requirements/requirements.txt
## common
#Flask-SQLAlchemy==2.4.1
#SQLAlchemy==1.3.15
#SQLAlchemy-JSONField==0.9.0
#SQLAlchemy-Utils==0.36.3
coverage==5.1
celery==4.4.2
pypandoc==1.5

pendulum==2.1.2


## Spark
pyspark==3.4.1
#py4j==0.10.9

## Azure
azure-mgmt-resource==18.0.0
azure-identity==1.6.0
azure-keyvault-secrets

# New
psycopg2
flake8
pandas
pytest

# deltalake
delta-spark==2.4.0

# Airflow
apache-airflow==2.6.3
apache-airflow-providers-mysql==5.2.0
apache-airflow-providers-postgres==5.6.0

## gcp
google-cloud-pubsub
google-cloud-storage
apache-beam

## aws
boto3==1.14.28